Refactoring
Set app name to source
Add guava
