+++ src/main/java/ca/bc/gov/splunknotificationservice/model/teams/TeamsFact.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.splunknotificationservice.Model.Teams;
+package ca.bc.gov.splunknotificationservice.model.teams;
 
 public class TeamsFact {
   private String name;
