+++ src/main/java/ca/bc/gov/splunknotificationservice/model/teams/TeamsMessage.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.splunknotificationservice.Model.Teams;
+package ca.bc.gov.splunknotificationservice.model.teams;
 
 public class TeamsMessage {
     private String text;
