+++ src/main/resources/application.properties
@@ -3,12 +3,3 @@ server.port=${PORT}
 
 splunk.teams_url=${TEAMS_URL}
 splunk.rocket_url=${ROCKET_URL}
-
-splunk.rocket_message_alias=POC Error
-splunk.teams_message_summary=Message From Splunk
-
-splunk.teams_enabled=false
-splunk.rocket_enabled=true
-splunk.teams_cards_enabled=false
-
-splunk.result_include_raw=true
